requirements.txt
Pillow
newspaper3k
urllib3
streamlit
beautifulsoup4
protobuf==3.20.0
jinja2>=3.0
   

streamlit
Pillow
beautifulsoup4
newspaper3k
nltk
transformers==4.35.2
datasets
requests
lxml[html_clean]
keras==2.11
torch==2.1.2


